+++ build-system/tasks/serve.js
@@ -25,7 +25,8 @@ var nodemon = require('nodemon');
 function serve() {
   nodemon({
     script: require.resolve('../server.js'),
-    watch: require.resolve('../server.js'),
+    watch: [require.resolve('../middleware-routing.js'),
+        require.resolve('../server.js')],
     env: {'NODE_ENV': 'development'},
   });
